+++ data/airport-coordinates.json.d.ts
@@ -1,5 +1,8 @@
 // Type definitions for airport-coordinates.json
 
+/**
+ * code, name, lat, lon
+ */
 export type AeroflyAirportCoordinates = [string, string, number, number];
 
 declare const AeroflyAirportCoordinatesList: AeroflyAirportCoordinates[];